refactor: move DependenciesGraph.merging to DependenciesGraph.swift
